move task related helpers to tasks.ts
